Implement comments field in rpc calls. The database stores an optional comment with every transaction.
This is implemented with an additional field to wtxmgr.TxRecord
called 'comment'. When serialized, TxRecord is the same except
with the comment added to the end as a JSON string. This
implementation is compatible with earlier serializations of
wtxmgr.TxRecord that may still be in the database.

glide.lock updated